lowpan: New LoWPAN CL architecture.
This rewrite allows to cope with connection to more then one node. The frames
are getting multiplexed in the converngence layer and directed to the different
connection objects. Lowpanstream takes care of breaking up bundles into several
segments if needed.

The multiplexing also makes room for the upcoming discovery frames.